Create Lat Long from Address
Enables user in Google sheet to highlight an address and the two cells next to the address, run the geocoder script, and then generate lat/long for the specified address. These coordinates are then retrieved from another function in the sheet and used to map the location of the address.